apps/api/pyproject.toml
[project]
name = "meridian-api"
version = "0.1.0"
description = "Meridian RAG API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    # AI/ML
    "openai>=1.30.0",
    "pinecone-client>=3.2.0",
    "tiktoken>=0.7.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.2.0",
    # Document processing
    "unstructured[pdf,doc]>=0.14.0",
    "pypdf>=4.2.0",
    # Database
    "sqlalchemy>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    # Redis
    "redis>=5.0.0",
    # Auth
    "python-jose[cryptography]>=3.3.0",
    # Utils
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.4",
    "mypy>=1.10.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
testpaths = ["tests"]
